+++ src/main/java/projet/rest/data/repositories/ConfirmationTokenRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.rest.data.models.ConfirmationToken;
 
-public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Double> {
+public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Long> {
 	ConfirmationToken findByConfirmationToken(String confirmationToken);
 }
